# Optionally use installed sdbus-cpp library