--- a/scripts/build-with-docker.sh
+++ b/scripts/build-with-docker.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
-docker build -t anbox/anbox-builder .
+DOCKER_BUILDKIT=1 docker build -t anbox/anbox-builder .
 docker run --rm --user=$(id -u):$(id -g) --volume=$PWD:/anbox anbox/anbox-builder /anbox/scripts/build.sh $@